Add balloon pods to reserve VM burst capacity (#103)
These are low-priority pods that are evicted quickly when workload pods
need the capacity. After they're evicted, they're re-scheduled in a
Pending state, which causes cluster autoscaler to provision new nodes.

The balloon pods also use the sim-runner image, which serves to pre-pull
the images onto newly provisioned nodes. This cuts down the observed
latency on new deployments.

This change adds 1 balloon pod in dev and 4 in prod, but this is easily
configurable.

Fixes: ELO-66

Signed-off-by: Akhil Velagapudi <4@4khil.com>