feat(sim-agent): add deriv to build sandbox vm (#200)
Bundling everything into an initramfs was a failed experiment. It's just
too slow to boot. Instead, use the "root=" kernel arguemnt to mount a
read-only squashfs as root and jump to its init process. This approach
is quite fast (boots, runs init process, and terminates in ~120ms). We
also get a compressed and immutable root filesystem as a bonus! Really
the main reason for using squashfs was just that it's easy to create the
filesystem as a non-root user (unlike ext4).

The squashfs file will likely always reside in host memory (via emptyDir
or tmpfs). So, we shouldn't pay a disk penalty.

Signed-off-by: Akhil Velagapudi <4@4khil.com>