infra: build Linux wheels on 22.04 (#654)
The problem with building on 20.04 is that it tricks auditwheel into
thinking that the wheel only depends on glibc 2.31 and it propagates
this information to PyPi. This allows users to `pip install elodin` on
Ubuntu 20.04 (or other older glibc distros) without any error. This is
bad because the XLA static library is compiled against glibc 2.35 and
will fail at runtime.

By building on 22.04, the glibc 2.35 symbols will be obvious and
auditwheel will report the dependency correctly.

Signed-off-by: Akhil Velagapudi <4@4khil.com>