fix(nox): add uv venv support for pyo3 (#702)
* fix(nox): add uv venv support for pyo3

Currently, pyo3 doesn't set up its module search path correctly when
insite a uv venv. I tried a whole bunch of things, but they all require
additional local setup. This is the only solutions that "just works"^TM
and lets you run `cargo test` in a venv with uv-managed Python.

It adds the system and venv site-packages to the modules search path on
initialization. This is only needed when doing Rust -> Python. The search
paths are setup fine when doing Python -> Rust -> Python. Turns out we
only do Rust -> Python in the tests for nox.

Signed-off-by: Akhil Velagapudi <4@4khil.com>

* fix: linux build

Signed-off-by: Akhil Velagapudi <4@4khil.com>

---------

Signed-off-by: Akhil Velagapudi <4@4khil.com>